heatmaps for reports
reports heatmaps down to upazila level
